[feat] 챗봇 activity 생성
